Add a flag to enable RBAC per tenant
This adds the flag `enforcement_enabled` to the `Tenant` model in the public
schema, which will serve as a way to enable or disable RBAC per tenant.

This flag will be used to allow global access and return all roles for requests
when enabled. By default, this flag is disabled.

Following PRs will make use of this, and implement an endpoint to update it.

- add a boolean field to the Tenant model
- add a migration file

In order to get this up locally, run:

```
./rbac/manage.py makemigrations
make run-migrations
```

You should then be able to see the column in the tenant table.